Isolate session manager types
